Change to stow layout
This commit has breaking changes for the previous setups!

Stow might be the better choice for installing the dotfiles. The
structure has been changed to be compatible with GNU stow.

Also added new Kitty config. It has no repo of it's own (yet), because
it currently consists of a single file.